move context mock up in directory tree
